+++ WebApiServer/Controllers/OwnerController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Service.Dto;
 using Service.Interfaces;
+using System.Security.Claims;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -17,38 +19,67 @@ namespace WebApiServer.Controllers
         }
         // GET: api/<OwnerController>
         [HttpGet]
-        public List<OwnerDto> Get()
+
+        public async Task<List<OwnerDto>> Get()
         {
-            return service.GetAll();
+            return await service.GetAll();
         }
 
         // GET api/<OwnerController>/5
         [HttpGet("{id}")]
-        public OwnerDto Get(int id)
+
+		public async Task<OwnerDto> Get(int id)
         {
-            return service.GetById(id);
+            return await service.GetById(id);
         }
 
         // POST api/<OwnerController>
         [HttpPost]
-        public async Task<IActionResult> Post([FromForm] OwnerDto value)
+        [Authorize]
+        public async Task<OwnerDto> Post([FromForm] OwnerDto value)
         {
-            if (value.FileImage != null)
-            {
-                var path = Path.Combine(Environment.CurrentDirectory, "Images/", value.FileImage.FileName);
-                using (FileStream fs = new FileStream(path, FileMode.Create))
-                {
-                    value.FileImage.CopyTo(fs);
-                    fs.Close();
-                }
-            }
-            return await service.AddItem(value);
+			if (value.FileImage != null)
+			{
+				var folderPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images");
+				if (!Directory.Exists(folderPath))
+					Directory.CreateDirectory(folderPath);
+
+				var fileName = Guid.NewGuid().ToString() + "_" + value.FileImage.FileName;
+				var fullPath = Path.Combine(folderPath, fileName);
+
+				using (var fs = new FileStream(fullPath, FileMode.Create))
+				{
+					await value.FileImage.CopyToAsync(fs); // שימוש ב-Async
+				}
+				var urlForClient = "/Images/" + fileName;
+				value.Image = urlForClient;
+			}
+			return await service.AddItem(value);
         }
 
         // PUT api/<OwnerController>/5
         [HttpPut("{id}")]
-        async public Task<IActionResult> Put(int id, [FromForm] OwnerDto value)
+        [Authorize(Roles = "owner,admin")]
+        async public Task Put(int id, [FromForm] OwnerDto value)
         {
+            var currentUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            var currentUserRole = User.FindFirst(ClaimTypes.Role)?.Value;
+            var owner = await service.GetById(id);
+
+            if (owner == null)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
+
+            if (currentUserRole == "owner")
+            {
+                if (owner.UserId != currentUserId)
+                {
+                    Response.StatusCode = 403;
+                    return;
+                }
+            }
             if (value.FileImage != null)
             {
                 var folderPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images");
@@ -63,15 +94,36 @@ namespace WebApiServer.Controllers
                     await value.FileImage.CopyToAsync(fs); // שימוש ב-Async
                 }
                 var urlForClient = "/Images/" + fileName;
-            }
-            return await service.UpdateItem(id, value);
+				value.Image = urlForClient;
+			}
+            await service.UpdateItem(id, value);
+
         }
 
         // DELETE api/<OwnerController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        [Authorize(Roles = "owner,admin")]
+        public async Task Delete(int id)
         {
-            service.DeleteItem(id);
+			var currentUserId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+			var currentUserRole = User.FindFirst(ClaimTypes.Role)?.Value;
+			var owner = await service.GetById(id);
+
+			if (owner == null)
+			{
+				Response.StatusCode = 404;
+				return;
+			}
+
+			if (currentUserRole == "owner")
+			{
+				if (owner.UserId != currentUserId)
+				{
+					Response.StatusCode = 403;
+					return;
+				}
+			}
+			await service.DeleteItem(id);
         }
     }
 }
